Moving UI rendering to new ui::render module and creating ui::widget module for widget-building
