+++ src/app.rs
@@ -100,7 +100,7 @@ impl App {
             if let Ok(size) = terminal.backend().size() {
                 state.set_terminal_size(size);
             };
-            terminal.draw(|frame| crate::ui::render::all(frame, &state))?;
+            terminal.draw(|frame| crate::ui::render(frame, &state))?;
             if !terminal_event_handler.handle_next(&mut state)? {
                 debug!("Received application exit request.");
                 break;
